recogniseCdf {goftest}	R Documentation

Explanatory Name of Distribution Function

Description

Recognises many standard cumulative distribution functions and returns a string describing the distribution.

Usage

recogniseCdf(s="punif")

Arguments

s
A single character string giving the name of an R function that calculates cumulative probabilities.

Details

The list of recognised distribution functions includes all those available in the stats package and in goftest.

By convention, the name of a cumulative distribution function begins with the letter p. For example, punif is the cumulative distribution function of the uniform distribution.

The initial letter p can be omitted in the function recogniseCdf.

Value

Character string, or NULL if the name is not recognised.

Author(s)

Adrian Baddeley.

See Also

pAD

Examples

   recogniseCdf("punif")
   recogniseCdf("unif")
   recogniseCdf("pt")

[Package goftest version 1.2-3 Index]